EDW-101: Solved vercel building issue.
