+++ lib/cmsClient.js
@@ -1,23 +1,5 @@
 import axios from "axios";
-
-async function login() {
-  const reqAuth = await axios.post(
-    process.env.BACKEND_URL + "/api/auth/local",
-    {
-      identifier: process.env.WEBSITE_USER,
-      password: process.env.WEBSITE_USER_PASSWORD,
-    }
-  );
-  return reqAuth.data.jwt;
-}
-
-export async function loginOnClient() {
-  const reqAuth = await axios.post(process.env.backendUrl + "/api/auth/local", {
-    identifier: process.env.web_user,
-    password: process.env.web_code,
-  });
-  return reqAuth.data.jwt;
-}
+import { getAuthToken } from "./auth";
 
 export async function getDepartments() {
   try {
@@ -50,7 +32,7 @@ export async function getDepartments() {
 
 export async function getBlogs() {
   try {
-    const token = await login();
+    const token = await getAuthToken();
 
     const res = await axios({
       method: "GET",
@@ -62,12 +44,9 @@ export async function getBlogs() {
       },
     });
 
-    const blogs = await res.data.data;
-
-    return blogs;
+    return res.data.data || [];
   } catch (err) {
-    //console.log(err);
-    return false;
+    return [];
   }
 }
 
@@ -148,7 +127,7 @@ export const getFeaturedProjects = async () => {
 
 export const getProjects = async () => {
   try {
-    let token = await login();
+    let token = await getAuthToken();
     let projects = await axios({
       method: "GET",
       url: "https://cms.educaciondiversa.com/api/projects?populate=*,mainPicture,project_category,department.picture,project_template,campaignPicture,pictures,sustainable_dev_goals.picture,resources,contentDownload",
@@ -167,7 +146,7 @@ export const getProjects = async () => {
 
 export const getAboutUsPageData = async () => {
   try {
-    let token = await login();
+    let token = await getAuthToken();
     let pageData = await axios({
       method: "GET",
       url: "https://cms.educaciondiversa.com/api/about-us-page?populate=*",
@@ -186,7 +165,7 @@ export const getAboutUsPageData = async () => {
 
 export const getMembers = async () => {
   try {
-    let token = await login();
+    let token = await getAuthToken();
     let members = await axios({
       method: "GET",
       url: "https://cms.educaciondiversa.com/api/members?populate=*",
